Refactor API service and authentication tools
- Removed the old simpleApiService and replaced it with a new structure for handling API calls.
- Introduced a token manager for Azure authentication, supporting interactive device code flow.
- Updated the apiCallTool to support various authentication methods including bearer, basic, and interactive.
- Enhanced error handling and response formatting in API call tools.
- Removed deprecated tools related to OpenAPI specifications and generated API calls.
- Added tests for interactive authentication to ensure functionality.
- Updated import paths to reflect the new service structure.
