--- api-tools/docs/project-summary.md
@@ -1,63 +1,0 @@
-# Completed Changes for API Tools MCP Server
-
-## Overview of Completed Tasks
-
-1. Successfully rebranded the project from "PDF Tools" to "API Tools"
-2. Added new API integration functionality alongside existing PDF tools
-3. Implemented authentication methods for API calls
-4. Added OpenAPI specification exploration capabilities
-5. Created tools for generating API call templates
-6. Updated documentation to reflect new features
-7. Added tests for API functionality
-
-## Implemented API Tools
-
-1. **call_api**: Makes authenticated API calls to a specified endpoint
-   - Supports multiple HTTP methods (GET, POST, PUT, PATCH, DELETE)
-   - Includes support for authentication (Bearer token, Basic auth)
-   - Handles query parameters, headers, and request bodies
-
-2. **get_api_operations**: Parses OpenAPI specifications
-   - Retrieves available operations from an API
-   - Can filter operations by operation ID
-   - Supports local and remote API specifications
-
-3. **generate_api_call**: Creates templates for API calls
-   - Uses OpenAPI specifications to build accurate request templates
-   - Populates parameters, headers, and request bodies
-   - Handles complex schema definitions
-
-## Authentication Methods
-
-- **Bearer token authentication**: For token-based APIs (OAuth 2.0, JWT)
-- **Basic authentication**: For username/password authentication
-- **No authentication**: For public APIs
-
-## Testing and Verification
-
-- Created test script for API call functionality
-- Verified API calls to public endpoints
-- Tested authentication methods
-
-## Documentation
-
-- Updated README.md with new API tools information
-- Created detailed API integration documentation
-- Added example usage for common API scenarios
-
-## Future Enhancements
-
-1. **Additional Authentication Methods**:
-   - OAuth 2.0 flow support
-   - API key authentication
-   - Custom authentication schemes
-
-2. **Extended API Tools**:
-   - WebSocket support
-   - GraphQL integration
-   - API testing and validation
-
-3. **Integration with Identity Services**:
-   - Azure AD integration
-   - Microsoft Identity Platform support
-   - Managed identities for cloud deployments
